refactor(shop): 테스트 코드 구조 개선 및 설정 정리
- 테스트 파일을 public/private으로 분리하여 인증 테스트
- Playwright 설정 파일 정리 및 프로젝트별 baseURL 설정
- 테스트 실행 스크립트를 admin/shop으로 분리
