+++ package.json
@@ -8,7 +8,8 @@
     "dev": "turbo dev",
     "dev:shop": "turbo dev --filter=shop",
     "dev:admin": "turbo dev --filter=admin",
-    "test": "exec playwright test",
+    "test:admin": "exec playwright test --project=admin --project=admin-setup",
+    "test:shop": "exec playwright test --project=shop --project=shop-setup",
     "prepare": "husky",
     "format": "biome format --write",
     "lint": "biome lint",
